Poprawki zad 9

+++ src/creature/Feedable.java
@@ -7,7 +7,7 @@ public interface Feedable {
 
 
 
-    void feed(int foodWeight);
+    void feed(Double foodWeight);
 
 
 }
